pyproject.toml
[tool.poetry]
name = "lego"
version = "1.0.0"
description = "Backend for abakus.no"
authors = ["Webkom <webkom@abakus.no>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.16"
ipython = "8.29.0"
psycopg2-binary = "2.9.6"
unicode-slugify = "0.1.5"
pyyaml = "6.0.1"
redis = "4.5.5"
hiredis = "2.2.3"
certifi = "2024.7.4"
elasticsearch = "7.16.2"
celery = ">=5.4,<6"
stripe = "3.4.0"
boto3 = "1.28.1"
beautifulsoup4 = "4.12.2"
bleach = "5.0.1"
requests = "2.32.2"
ldap3 = "2.9.1"
google-api-python-client = "^2.155.0"
google-auth = "^2.37.0"
libthumbor = "2.0.2"
django-health-check = "3.17.0"
djangorestframework = "3.14.0"
drf-jwt = "1.19.2"
djangorestframework-camel-case = "1.4.2"
django-extensions = "3.2.3"
django-autoslug = "1.9.9"
django-oauth-toolkit = "1.6.3"
django-cors-headers = "4.5.0"
django-mptt = "0.14.0"
django-filter = "23.2"
django-environ = "0.10.0"
django-redis = "5.4.0"
django-ipware = "5.0.0"
django-phonenumber-field = { extras = ["phonenumberslite"], version = "7.1.0" }
django-push-notifications = "3.0.0"
#
# Logging/errors
structlog = "21.5.0"
structlog-sentry = "1.4.0"
sentry-sdk = "1.25.0"
analytics-python = "1.2.9"
#
# channels
channels = { version = "4.0.0", extras = ["daphne"] }
channels_redis = "4.1.0"
# Implicit dependency
flatbuffers = "23.5.26"
#
# Ical-export
django-ical = "1.9.1"
icalendar = "5.0.7"
#
# Email
aiosmtpd = "1.4.6"
premailer = "3.10.0"
#
# Docs
Pygments = "2.15.1"
Markdown = "3.4.3"
coreapi = "2.3.3"
authlib = "^1.3.1"
weasyprint = "^63.0"
matplotlib = "^3.9.2"
django-expo-notifications = "^0.7.0"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "4.1.0"
tox = "4.11.3"
tblib = "^1.7.0"

[tool.poetry.group.coverage.dependencies]
coverage = "7.4.1"

[tool.poetry.group.mypy.dependencies]
mypy = "1.5.0"
mypy-extensions = "1.0.0"
django-stubs = "4.2.1"
djangorestframework-stubs = "3.14.0"
types-bleach = "6.1.0.0"
types-certifi = "2021.10.8.3"
types-docutils = "0.20.0.1"
types-Markdown = "3.4.2.9"
types-python-slugify = "8.0.0.3"
types-PyYAML = "6.0.12.10"
types-requests = "2.31.0.1"
types-six = "1.16.21.8"
types-urllib3 = "1.26.25.13"
types-python-dateutil = "2.8.19.13"

[tool.poetry.group.formatting.dependencies]
black = "24.10.0"
isort = "5.12.0"

[tool.poetry.group.flake8.dependencies]
flake8 = "6.0.0"
flake8-bugbear = "23.3.23"
flake8-comprehensions = "3.10.1"
flake8-print = "^5.0.0"

[tool.poetry.group.docs.dependencies]
sphinx = "6.2.1"
sphinx_rtd_theme = "1.2.1"

[tool.poetry.group.prod]
optional = true

[tool.poetry.group.prod.dependencies]
uwsgi = "2.0.22"
flower = "1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.django-stubs]
django_settings_module = "lego.settings"


[tool.isort]
profile = "black"
skip_gitignore = true
default_section = "THIRDPARTY"
known_first_party = "lego"
known_django = ["django", "rest_framework", "django_filters"]
sections = [
    "FUTURE",
    "STDLIB",
    "DJANGO",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
]
combine_as_imports = true


[tool.flake8]
config_file = ".flake8"


# [tool.black]
# target-version = ['py311']
# line-length = 88
